requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==2.1.4
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
comtypes==1.1.11
cryptography==36.0.2
cssselect==1.1.0
cssutils==2.4.0
desktop==0.4.2
DingtalkChatbot==1.5.3
docopt==0.6.2
et-xmlfile==1.1.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
jdcal==1.4.1
lxml==4.8.0
openpyxl==2.4.11
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
premailer==3.10.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.1.1
pytest-html==3.1.1
pytest-metadata==2.0.1
pywin32==303
pywinauto==0.6.8
PyYAML==6.0
records==0.5.3
requests==2.27.1
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
tablib==3.2.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
yagmail==0.15.277
